refactor: restructure notification services and introduce dependency injection
